feat: 流水线插件支持展示获得的荣誉和SLA等信息 #8114
