+++ src/backend/ci/core/store/biz-store/src/main/kotlin/com/tencent/devops/store/service/common/StoreCommonService.kt
@@ -30,7 +30,6 @@ package com.tencent.devops.store.service.common
 import com.tencent.devops.store.pojo.common.ReleaseProcessItem
 import com.tencent.devops.store.pojo.common.StoreProcessInfo
 import com.tencent.devops.store.pojo.common.StoreShowVersionInfo
-import com.tencent.devops.store.pojo.common.enums.ErrorCodeTypeEnum
 import com.tencent.devops.store.pojo.common.enums.ReleaseTypeEnum
 import com.tencent.devops.store.pojo.common.enums.StoreTypeEnum
 import org.jooq.DSLContext
@@ -130,14 +129,4 @@ interface StoreCommonService {
         releaseType: ReleaseTypeEnum?,
         version: String?
     ): StoreShowVersionInfo
-
-    /**
-     * 判断错误码是否合规
-     */
-    fun isComplianceErrorCode(
-        storeCode: String,
-        storeType: StoreTypeEnum,
-        errorCode: Int,
-        errorCodeType: ErrorCodeTypeEnum
-    ): Boolean
 }
