+++ src/backend/ci/core/store/biz-store/src/main/kotlin/com/tencent/devops/store/service/common/StoreErrorCodeService.kt
@@ -29,6 +29,8 @@ package com.tencent.devops.store.service.common
 
 import com.tencent.devops.common.api.pojo.Result
 import com.tencent.devops.store.pojo.common.ErrorCodeInfo
+import com.tencent.devops.store.pojo.common.enums.ErrorCodeTypeEnum
+import com.tencent.devops.store.pojo.common.enums.StoreTypeEnum
 
 interface StoreErrorCodeService {
 
@@ -39,4 +41,14 @@ interface StoreErrorCodeService {
         userId: String,
         errorCodeInfo: ErrorCodeInfo
     ): Result<Boolean>
+
+    /**
+     * 判断错误码是否合规
+     */
+    fun isComplianceErrorCode(
+        storeCode: String,
+        storeType: StoreTypeEnum,
+        errorCode: Int,
+        errorCodeType: ErrorCodeTypeEnum
+    ): Boolean
 }
