+++ src/backend/ci/core/store/biz-store/src/main/kotlin/com/tencent/devops/store/service/common/impl/StoreIndexManageServiceImpl.kt
@@ -27,7 +27,6 @@
 
 package com.tencent.devops.store.service.common.impl
 
-import com.tencent.bkrepo.common.api.exception.ErrorCodeException
 import com.tencent.devops.common.api.constant.CommonMessageCode
 import com.tencent.devops.common.api.constant.CommonMessageCode.ERROR_INVALID_PARAM_
 import com.tencent.devops.common.api.exception.ErrorCodeException
